# openapi.yaml
openapi: 3.1.0
servers:
  - url: https://backend.dm.teamofsilicons.com
info: {title: DM API, version: 0.10.0}
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: Exchange an organization-bound IAM SLT
      security: []
      parameters:
        - {$ref: '#/components/parameters/IdempotencyKey'}
        - {$ref: '#/components/parameters/TestingKey'}
      description: Maximum JSON body 16 KiB. No separate Bearer/X-Org-ID header required. Repeat the exact body with the same Idempotency-Key after an uncertain
        result. Credentials and backend app secrets must never be logged. The returned session is live-verified by IAM and has no-store/no-cache response
        headers.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [type, data]
              properties:
                type: {type: string, const: login}
                data: {$ref: '#/components/schemas/LoginInput'}
      responses:
        '200':
          description: Verified application session
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [type, data]
                properties:
                  type: {type: string, const: login}
                  data: {$ref: '#/components/schemas/ApplicationSession'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        default: {$ref: '#/components/responses/Error'}
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      operationId: refreshSession
      summary: Rotate a DM application refresh token
      security: []
      parameters:
        - {$ref: '#/components/parameters/IdempotencyKey'}
        - {$ref: '#/components/parameters/TestingKey'}
      description: Maximum JSON body 16 KiB. No separate Bearer/X-Org-ID header required. Repeat the exact body with the same Idempotency-Key after an uncertain
        result. Credentials and backend app secrets must never be logged. The returned session is live-verified by IAM and has no-store/no-cache response
        headers.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [type, data]
              properties:
                type: {type: string, const: refresh}
                data: {$ref: '#/components/schemas/RefreshInput'}
      responses:
        '200':
          description: Verified application session
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [type, data]
                properties:
                  type: {type: string, const: refresh}
                  data: {$ref: '#/components/schemas/ApplicationSession'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        default: {$ref: '#/components/responses/Error'}
  /api/v1/auth/logout:
    post:
      tags: [Auth]
      operationId: logout
      summary: Revoke one application token or refresh family
      security: []
      parameters:
        - {$ref: '#/components/parameters/IdempotencyKey'}
        - {$ref: '#/components/parameters/TestingKey'}
      description: Maximum JSON body 16 KiB. No separate Bearer/X-Org-ID header required. Repeat the exact body with the same Idempotency-Key after an uncertain
        result. Credentials and backend app secrets must never be logged. Use the current refresh token for full-family logout; an access token revokes
        only itself.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [type, data]
              properties:
                type: {type: string, const: logout}
                data: {$ref: '#/components/schemas/LogoutInput'}
      responses:
        '204': {description: Token revoked}
        '401': {$ref: '#/components/responses/Unauthorized'}
        default: {$ref: '#/components/responses/Error'}
...
components:
  securitySchemes:
    honeycombService: {type: http, scheme: bearer, description: 'Dedicated Honeycomb service credential, independent of test sessions.'}
    bearerAuth: {type: http, scheme: bearer, bearerFormat: IAM application access token (oat_)}
    testingEnvironmentKey:
      type: apiKey
      in: header
      name: X-Testing-Environment-Key
      description: 32-character DM testing root key. Selects an isolated data plane; normal actions still require its IAM user session.
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, minLength: 8, maxLength: 255, pattern: '^[!-~]{8,255}$'}
...
    TestingKey:
      name: X-Testing-Environment-Key
      in: header
      required: false
      schema: {type: string, pattern: '^(ask_[A-Za-z0-9_-]{43}|[A-Za-z0-9]{32})$'}
      description: IAM test app secret for automatic discovery, or a legacy DM root key. Omit for production. Never send the IAM environment root key.
...
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            type: object
            additionalProperties: false
            required: [type, data]
            properties:
              type: {type: string, const: error}
              data: {$ref: '#/components/schemas/Error'}
    Unauthorized:
      description: Authentication credentials are missing or invalid
      content:
        application/json:
          schema:
            type: object
            additionalProperties: false
            required: [type, data]
            properties:
              type: {type: string, const: error}
              data: {$ref: '#/components/schemas/Error'}
  schemas:
...
    MemberRef:
      type: object
      required: [type, id]
      properties:
        type:
          type: string
          enum: [carbon, silicon]
        id:
          type: string
          description: IAM account ID.
          examples: [alice@example.com, 'assistant:acme']
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: {type: string}
            message: {type: string}
    ApplicationSession:
      type: object
      properties:
        access_token: {type: string, writeOnly: false}
        refresh_token: {type: string}
        token_type: {const: Bearer}
        expires_in: {type: integer, minimum: 1}
        scope: {type: string}
        member: {$ref: '#/components/schemas/MemberRef'}
        organization_id: {type: string, description: Initial workspace; first currently selected organization handle in sorted order.}
        organization_ids:
          type: array
          minItems: 1
          uniqueItems: true
          description: Organizations explicitly selected in IAM with current active memberships. Each request still authorizes X-Org-ID.
          items: {type: string}
      required: [access_token, refresh_token, token_type, expires_in, scope, member, organization_id, organization_ids]
...
    LoginInput:
      type: object
      properties:
        slt: {type: string, writeOnly: true, description: Opaque IAM short-lived token. Current IAM wire values use oac_; the field is named slt. Never
            assume the field name is a token prefix.}
      required: [slt]
      additionalProperties: false
    RefreshInput:
      type: object
      properties:
        refresh_token: {type: string, writeOnly: true, pattern: ^ort_}
      required: [refresh_token]
      additionalProperties: false
    LogoutInput:
      type: object
      properties:
        token: {type: string, writeOnly: true}
      required: [token]
      additionalProperties: false
...